create op to materialize mongo alldocs collection
